fix some bugs and improve dosomething()
fix some bugs
in dosomething() add printing of total bytes copied
