+++ src/kopyx.h
@@ -1,24 +1,9 @@
 #ifndef KOPY_H_
 #define KOPY_H_
 
-//#define _GNU_SOURCE
-#define _XOPEN_SOURCE 500
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <stdbool.h>
-#include <string.h>
-#include <time.h>
-#include <ctype.h>
-#include <fcntl.h>
-#include <unistd.h>
-#include <sys/statvfs.h>
-#include <sys/sendfile.h>
-#include <errno.h>
-#include <sys/types.h>
+#define _DEFAULT_SOURCE
 #include <dirent.h>
 #include <glob.h>
-#include <getopt.h>
 
 #include "ui_funcs.h"
 #include "os_funcs.h"
@@ -29,19 +14,15 @@
 #define DBG_MSG(msg, ...) {}
 #endif
 
+/* look for the file in the specified directory */
 bool kopyx(const char *fromdir);
+/* perform glob in case of wildcard */
 void doglob(const char *fullpath);
+/* does what the user specified to do on the found file */
 void dosomething(const char *source);
-bool diskspace(const char *source, const char *dest);
-void deletefile(const char *fname);
-void showtoscreen(const char *from);
-void find(const char *fname);
 
 extern bool wildcard, found_one, delfile, find_only, verify, standardoutput, info, include_subdirs, noconfirm;
 extern char *pattern, *todir;
 extern ino_t todir_inode;
-
-#define NSECT 64
-#define BYTES 512
 
 #endif // KOPY_H
